signficant cleanup of repo
