+++ .github/workflows/tests.yml
@@ -7,6 +7,7 @@ on:
     branches:
       - "*"
 jobs:
+  # We don't currently have any tests that onyl run with JS. perhaps in the future.
   # js-test:
   #   runs-on: ubuntu-latest
   #   strategy:
@@ -51,6 +52,10 @@ jobs:
     - uses: microsoft/playwright-github-action@v1
     - name: Setup docker
       run: cd tests/rails && docker-compose up -d
+
+    - name: Set parallel workers env variable
+      run: echo "PARALLEL_WORKERS=$(expr $(nproc) \* 6)" >> $GITHUB_ENV
+
     - uses: ruby/setup-ruby@v1
       with:
         ruby-version: ${{ matrix.ruby }}
