+++ package.json
@@ -7,6 +7,7 @@
   "main": "dist/bundle/index.common.js",
   "module": "dist/bundle/index.module.js",
   "scripts": {
+  	"setup": "pnpm install && cd tests/rails && pnpm install && bundle install",
     "build": "rm -rf dist && pnpm build:types && node esbuild.config.js",
     "build:types": "tsc",
     "lint": "ts-standard src/",
