+++ src/elements/trix.ts
@@ -5,16 +5,27 @@ import StarterKit from '@tiptap/starter-kit'
 import Link from "@tiptap/extension-link"
 import Focus from "@tiptap/extension-focus";
 import Placeholder from "@tiptap/extension-placeholder"
+import Attachment from 'src/extensions/attachment'
 
-import { css, CSSResult, html, LitElement, PropertyDeclarations, PropertyValueMap, TemplateResult } from 'lit'
+import { CSSResult, html, LitElement, PropertyDeclarations, PropertyValueMap, TemplateResult } from 'lit'
 import { ref, createRef, Ref } from 'lit/directives/ref.js';
+
+/** Imports <role-tooltip> and <role-toolbar> */
 import "role-components"
 
-import { AttachmentUpload, AttachmentManager } from "../attachment-upload"
-import Attachment from '../attachment'
-import * as icons from '../icons'
-import { normalize } from '../styles/normalize'
-import type { Maybe } from '../types'
+import { AttachmentUpload } from "src/models/attachment-upload"
+import { AttachmentManager } from "src/models/attachment-manager";
+
+/** This will go away in favor of lazy loaded SVGs. */
+import * as icons from 'src/views/icons'
+
+
+import { normalize } from 'src/styles/normalize';
+import trixStyles from 'src/styles/trix';
+
+import { TipTapAddAttachmentEvent } from "src/events/tip-tap-add-attachment-event";
+
+import type { Maybe } from 'src/types'
 
 export const isiOS = /Mac|iOS|iPhone|iPad|iPod/i.test(window.navigator.platform)
 
@@ -37,27 +48,14 @@ export const config = {
   linkDialogUnlink: `Unlink`
 }
 
-export class TipTapAddAttachmentEvent extends Event {
-  attachment: AttachmentManager
-
-  static get eventName (): "tip-tap-add-attachment" {
-    return "tip-tap-add-attachment"
-  }
-
-  constructor (attachment: AttachmentManager, options: Partial<EventInit> = {}) {
-    if (options.bubbles == null) options.bubbles = true
-    if (options.composed == null) options.composed = true
-    if (options.cancelable == null) options.cancelable = true
-
-    super(TipTapAddAttachmentEvent.eventName, options);
-    this.attachment = attachment
-  }
-}
-
 export type ToolbarButton<T extends string> =
 | `button button__${T}`
 | `button button__${T} button--active`
 
+/**
+ * This is the meat and potatoes. This is the <tip-tap-trix> element you'll
+ *   see. This is what wraps everything into 1 coherent element.
+ */
 export class TipTapElement extends LitElement {
   readonly: boolean = false;
   linkInputRef: Ref<HTMLInputElement> = createRef()
@@ -77,234 +75,14 @@ export class TipTapElement extends LitElement {
     }
   }
 
-  static get styles (): CSSResult {
-    return css`
-      ${normalize}
-      ${tipTapCoreStyles}
-
-      :host {
-        display: block;
-        --border-color: #cecece;
-        --placeholder-text-color: #cecece;
-        --input-focus-ring: 0 0 2px 1px #005a9c;
-
-        --active-button-background-color: rgb(226 239 255);
-
-        --button-text-color: #889;
-        --button-border-color: #005a9c;
-        --button-background-color: hsl(219, 26%, 95%);
-
-        --disabled-button-text-color: #d1d5db;
-        --disabled-button-border-color: #d1d5db;
-        --disabled-button-background-color: #d1d5db;
-
-        --toolbar-text-color: hsl(219, 6%, 43%);
-
-        --link-dialog-border-color: #005a9c;
-        --link-dialog-background-color: hsla(219, 26%, 95%, 0.5);
-
-        --link-dialog-input-invalid-background-color: #ffdddd;
-        --link-dialog-input-invalid-border-color: red;
-
-        color: #374151;
-      }
-
-      img, svg {
-        width: 100%;
-      }
-
-      img, svg, figure {
-        max-width: 100%;
-        height: auto;
-        display: block;
-      }
-
-      figure, p {
-        padding: 0;
-        margin: 0;
-      }
-
-      figure {
-        position: relative;
-      }
-
-      .ProseMirror .placeholder {
-        position: absolute;
-        pointer-events: none;
-        color: var(--placeholder-text-color);
-        cursor: text;
-        content: "";
-      }
-
-      .ProseMirror {
-        border: 1px solid var(--border-color);
-        border-radius: 3px;
-        margin: 0;
-        padding: 0.4em 0.6em;
-        min-height: 200px;
-        outline: transparent;
-      }
-
-      .toolbar {
-        color: var(--toolbar-text-color);
-        overflow: auto;
-      }
-
-      .toolbar__button {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        min-height: 32px;
-        margin-right: -3px;
-        min-width: 32px;
-        position: relative;
-        border: 1px solid var(--border-color);
-        border-radius: 4px;
-      }
-
-      .toolbar__button[aria-disabled="true"] {
-        color: var(--disabled-button-text-color);
-        border-color: var(--disabled-button-border-color);
-      }
-
-      .toolbar__button[aria-disabled="true"]:focus {
-        border-color: var(--disabled-button-border-color);
-      }
-
-      .toolbar__button svg {
-        flex-shrink: 0;
-        height: 24px;
-        width: 24px;
-      }
-
-      button:is(:focus, :hover):not([aria-disabled="true"], :disabled) {
-        outline: transparent;
-        box-shadow: 0 0 0 1px var(--button-border-color);
-        border-color: var(--button-border-color);
-        background-color: var(--button-background-color);
-      }
-
-      .toolbar__button:is([aria-disabled="true"]:not([part~="button--active"])) {
-        color: var(--disabled-button-text-color);
-        border-color: var(--disabled-button-border-color);
-      }
-
-      .toolbar__button:is(:focus, :hover):is([aria-disabled="true"]:not([part~="button--active"])) {
-        outline: transparent;
-        color: var(--disabled-button-text-color);
-        border-color: var(--disabled-button-border-color);
-        box-shadow: 0 0 0 1px var(--disabled-button-border-color);
-      }
-
-      .toolbar__button:is([part~="button--active"]),
-      .toolbar__button:is([part~="button--active"]):is(:hover, :focus) {
-        background-color: var(--active-button-background-color);
-      }
-
-      .toolbar__button:is([part~="button__link"], [part~="button__orderedList"]) {
-        margin-right: 1rem;
-      }
-
-      .toolbar__button[part~="button__files"] {
-        margin-right: auto;
-      }
-
-      .link-dialog {
-        position: absolute;
-        z-index: 1;
-        height: 100%;
-        width: 100%;
-        padding: 1px;
-      }
-
-      .link-dialog__container {
-        display: flex;
-        align-items: center;
-        background: white;
-        box-shadow: 0 0.3em 1em #ccc;
-        max-width: 600px;
-        padding: 0.75rem 0.4rem;
-        border-radius: 8px;
-        border-top: 2px solid #ccc;
-      }
-
-      .link-dialog__input {
-        border: 1px solid #374151;
-        border-radius: 4px;
-        padding: 0.4em 0.6em;
-        flex: 1 1 auto;
-      }
-
-      .link-dialog__input:is(:focus) {
-        outline: transparent;
-        box-shadow: var(--input-focus-ring);
-        border-color: var(--link-dialog-border-color);
-        background-color: var(--link-dialog-background-color);
-      }
-
-      .link-validate:invalid {
-        outline: transparent;
-        background-color: var(--link-dialog-input-invalid-background-color);
-        border-color: var(--link-dialog-input-invalid-border-color);
-        box-shadow: none;
-      }
-
-      .link-dialog__button {
-        padding: 0.4em 0.6em;
-        border: 1px solid var(--button-border-color);
-        border-radius: 4px;
-      }
-
-      .link-dialog__buttons {
-        margin-right: 0.5em;
-        margin-left: 0.5em;
-      }
-
-      /* Attachments */
-      figure.has-focus {
-        outline: transparent;
-        box-shadow: 0 0 0 2px var(--button-border-color);
-      }
-
-      attachment-editor {
-        display: none;
-      }
-
-      .ProseMirror[contenteditable="true"] figure.has-focus attachment-editor {
-        display: flex;
-      }
-
-      figcaption {
-        position: relative;
-      }
-
-      .ProseMirror p.is-editor-empty:first-child::before,
-      figcaption p:first-child.is-empty::before {
-        color: #adb5bd;
-        content: attr(data-placeholder);
-      }
-
-      figcaption p:first-child.is-empty::before {
-        position: absolute;
-        top: 0;
-        left: 50%;
-        transform: translateX(-50%);
-        cursor: text;
-      }
-
-      .ProseMirror p.is-editor-empty:first-child::before {
-        float: left;
-        height: 0;
-        pointer-events: none;
-      }
-
-    `
+  static get styles (): CSSResult[] {
+  	return [normalize, tipTapCoreStyles, trixStyles]
   }
 
   connectedCallback (): void {
     super.connectedCallback()
 
-    addEventListener(TipTapAddAttachmentEvent.eventName, (event: TipTapAddAttachmentEvent) => {
+    document.addEventListener(TipTapAddAttachmentEvent.eventName, (event: TipTapAddAttachmentEvent) => {
       const { attachment, target } = event
 
       if (target instanceof HTMLElement && attachment.file) {
@@ -349,7 +127,10 @@ export class TipTapElement extends LitElement {
     // const div = document.createElement("div")
     // this.insertAdjacentElement("afterend", div)
     this.editor = this.setupEditor(element)
-    this.editorElement = element
+    this.editorElement = element.querySelector(".ProseMirror")
+
+    this.editorElement?.setAttribute('tabindex', '0');
+    this.editorElement?.setAttribute('role', 'textbox');
   }
 
 
@@ -377,6 +158,7 @@ export class TipTapElement extends LitElement {
   }
 
   setupEditor (element: Element): Editor {
+  	console.log(this.inputElement?.value)
     return new Editor({
       element,
       injectCSS: false,
@@ -827,12 +609,6 @@ export class TipTapElement extends LitElement {
         ${this.renderLinkCreationDialog()}
       </div>
     `
-  }
-}
-
-declare global {
-  interface WindowEventMap {
-    [TipTapAddAttachmentEvent.eventName]: TipTapAddAttachmentEvent
   }
 }
