+++ src/types.ts
@@ -1,1 +1,12 @@
+/* Utility Types */
 export type Maybe<T> = T | null | undefined
+
+/* Library specific types */
+export type AttachmentAttributes = {
+  file: File
+  src: string
+  sgid?: Maybe<string>
+  url?: Maybe<string>
+}
+
+
