+++ tests/rails/app/views/home/index.html.erb
@@ -1,5 +1,7 @@
 <br><br>
 
+<%= link_to "Posts", posts_path %>
+
 <h2>TipTap Editor</h2>
 <form>
   <input id="y" type="hidden" name="content">
