+++ tests/rails/app/views/posts/edit.html.erb
@@ -1,3 +1,4 @@
+<div style="max-width: 60ch; margin: 0 auto;">
 <h1>Editing post</h1>
 
 <%= render "form", post: @post %>
@@ -7,4 +8,5 @@
 <div>
   <%= link_to "Show this post", @post %> |
   <%= link_to "Back to posts", posts_path %>
+</div>
 </div>
