+++ tests/rails/app/views/posts/new.html.erb
@@ -1,3 +1,4 @@
+<div style="max-width: 60ch; margin: 0 auto;">
 <h1>New post</h1>
 
 <%= render "form", post: @post %>
@@ -7,3 +8,5 @@
 <div>
   <%= link_to "Back to posts", posts_path %>
 </div>
+
+<div style="max-width: 60ch; margin: 0 auto;">
